Vazby

+++ src/idea/idea.service.ts
@@ -1,57 +1,80 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, Logger } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { IdeaEntity } from 'src/idea/idea.entity';
+import { UserEntity } from 'src/user/user.entity';
 
 import { Repository } from 'typeorm';
-import { IdeaDTO } from './idea.dto';
+import { IdeaDTO, IdeaRO } from './idea.dto';
 
 @Injectable()
 export class IdeaService {
-    constructor(@InjectRepository(IdeaEntity)  
-    private ideaRepository: Repository<IdeaEntity>)
+    constructor(
+        @InjectRepository(IdeaEntity)  
+        private ideaRepository: Repository<IdeaEntity>,
+        @InjectRepository(UserEntity)
+        private userRepository: Repository<UserEntity>
+    )
     {}
 
-    showAll()
+    private toResponseObject(idea: IdeaEntity):IdeaRO
     {
-        return this.ideaRepository.find();
+        return {...idea, author:idea.author.toResponseObject(false)};
     }
 
-    async create(data: IdeaDTO)
+    private ensureOwnership(idea: IdeaEntity, userId: string)
     {
-        const idea = this.ideaRepository.create(data);
+        if(idea.author.id !== userId)
+        {
+            throw new HttpException("Incorrect user", HttpStatus.UNAUTHORIZED);
+        }
+    }
+
+    async showAll(): Promise<IdeaRO[]>
+    {
+        const ideas = await this.ideaRepository.find({relations:["author"]});
+        return ideas.map(idea => this.toResponseObject(idea));
+    }
+
+    async create(userId: string,data: IdeaDTO): Promise<IdeaRO>
+    {
+        const user = await this.userRepository.findOne({where:{id:userId}});
+        const idea = this.ideaRepository.create({...data, author:user});
         await this.ideaRepository.save(idea);
-        return idea;
+        return this.toResponseObject(idea);
     } 
 
-    async read(id: string){
-        const idea =  await this.ideaRepository.findOne({where: {id}});
+    async read(id: string): Promise<IdeaRO>{
+        const idea =  await this.ideaRepository.findOne({where: {id}, relations: ["author"]});
         if(!idea)
         {
             throw new HttpException("Not found!", HttpStatus.NOT_FOUND);
         }
-        return idea;
+        return this.toResponseObject(idea);
     }
 
-    async update(id:string, data: Partial<IdeaDTO>)
+    async update(id:string, userId: string, data: Partial<IdeaDTO>):Promise<IdeaRO>
     {
-        const idea = await this.ideaRepository.findOne({where: {id}});
+        let idea = await this.ideaRepository.findOne({where: {id}, relations:["author"]});
         if(!idea)
         {
             throw new HttpException("Not found!", HttpStatus.NOT_FOUND);
         }
+        this.ensureOwnership(idea, userId);
         await this.ideaRepository.update({id},data);
-        return await this.ideaRepository.findOne({id});
+        idea = await this.ideaRepository.findOne({where:{id},relations:["author"]});
+        return this.toResponseObject(idea);
     }
 
-    async destroy(id:string)
+    async destroy(id:string, userId: string)
     {
-        const idea = await this.ideaRepository.findOne({where: {id}});
+        const idea = await this.ideaRepository.findOne({where: {id}, relations:["author"]});
         if(!idea)
         {
             throw new HttpException("Not found!", HttpStatus.NOT_FOUND);
         }
+        this.ensureOwnership(idea, userId);
         await this.ideaRepository.delete({id});
-        return idea;
+        return this.toResponseObject(idea);
     }
 
     async deleteAll()
